+++ examples/plot_dipole.py
@@ -9,6 +9,8 @@ source field.
 ###############################################################################
 # First, import required modules
 import astropy.constants as const
+import astropy.units as u
+from astropy.coordinates import SkyCoord
 import matplotlib.pyplot as plt
 import matplotlib.patches as mpatch
 import numpy as np
@@ -78,15 +80,13 @@ fig, ax = plt.subplots()
 ax.set_aspect('equal')
 
 # Take 32 start points spaced equally in theta
-r = 1.01
-phi = np.pi / 2
-r = 1.01
-phi = np.pi / 2
-theta = np.linspace(0, np.pi, 33)
-x0 = np.array(coords.sph2cart(r, theta, phi)).T
+r = 1.01 * const.R_sun
+lon = np.pi / 2 * u.rad
+lat = np.linspace(-np.pi / 2, np.pi / 2, 33) * u.rad
+seeds = SkyCoord(lon, lat, r, frame=output.coordinate_frame)
 
 tracer = pfsspy.tracing.PythonTracer()
-field_lines = tracer.trace(x0, output)
+field_lines = tracer.trace(seeds, output)
 
 for field_line in field_lines:
     color = {0: 'black', -1: 'tab:blue', 1: 'tab:red'}.get(field_line.polarity)